# app/api/openapi.yaml
openapi: 3.0.0
info:
  title: AWS Services API
  version: 1.0.0
  description: Comprehensive API for managing AWS services including Lambda, API Gateway, IAM, SNS, SQS, and Step Functions
servers:
  - url: /api
    description: API v1
components:
  securitySchemes:
    AWS4-HMAC-SHA256:
      type: apiKey
      in: header
      name: Authorization
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  responses:
    BadRequestError:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnauthorizedError:
      description: AWS Authentication Failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        requestId:
          type: string
        timestamp:
          type: string

    # S3 Schemas
    S3Bucket:
      type: object
      properties:
        Name:
          type: string
        CreationDate:
          type: string
        Region:
          type: string
        Versioning:
          type: boolean
        Encryption:
          type: string
          enum: [AES256, aws:kms]

    # Lambda Schemas
    LambdaFunction:
      type: object
      properties:
        FunctionName:
          type: string
        Runtime:
          type: string
          enum: [nodejs18.x, python3.9, java11, dotnet6]
        Handler:
          type: string
        Role:
          type: string
        Code:
          type: object
          properties:
            ZipFile:
              type: string
            S3Bucket:
              type: string
            S3Key:
              type: string
        Environment:
          type: object
          properties:
            Variables:
              type: object
              additionalProperties:
                type: string
        MemorySize:
          type: integer
          minimum: 128
          maximum: 10240
        Timeout:
          type: integer
          minimum: 1
          maximum: 900
        Tags:
          type: object
          additionalProperties:
            type: string

    # API Gateway Schemas
    ApiGateway:
      type: object
      properties:
        Name:
          type: string
        Description:
          type: string
        EndpointType:
          type: string
          enum: [REGIONAL, EDGE, PRIVATE]
        Protocol:
          type: string
          enum: [HTTP, REST, WEBSOCKET]
        Stage:
          type: string
        Resources:
          type: array
          items:
            $ref: '#/components/schemas/ApiResource'
    ApiResource:
      type: object
      properties:
        Path:
          type: string
        Method:
          type: string
          enum: [GET, POST, PUT, DELETE, PATCH]
        Integration:
          type: object
          properties:
            Type:
              type: string
              enum: [AWS, HTTP, MOCK, AWS_PROXY, HTTP_PROXY]
            Uri:
              type: string

    # IAM Schemas
    IAMRole:
      type: object
      properties:
        RoleName:
          type: string
        Description:
          type: string
        AssumeRolePolicyDocument:
          type: object
        MaxSessionDuration:
          type: integer
        Path:
          type: string
        Tags:
          type: array
          items:
            type: object
            properties:
              Key:
                type: string
              Value:
                type: string
        ManagedPolicyArns:
          type: array
          items:
            type: string
        InlinePolicies:
          type: object
          additionalProperties:
            type: object

    # SNS Schemas
...
    DynamoDBTable:
      type: object
      required:
        - TableName
        - PartitionKey
      properties:
        TableName:
          type: string
          pattern: '^[a-zA-Z0-9_.-]+$'
          minLength: 3
          maxLength: 255
        PartitionKey:
          type: object
          required:
            - name
            - type
          properties:
            name:
              type: string
            type:
              type: string
              enum: [String, Number, Binary]
        SortKey:
          type: object
          required:
            - name
            - type
          properties:
            name:
              type: string
            type:
              type: string
              enum: [String, Number, Binary]
        BillingMode:
          type: string
          enum: [PROVISIONED, PAY_PER_REQUEST]
          default: PAY_PER_REQUEST
        ProvisionedThroughput:
          type: object
          properties:
            ReadCapacityUnits:
              type: integer
              minimum: 1
            WriteCapacityUnits:
              type: integer
              minimum: 1
        Tags:
          type: array
          items:
            type: object
            properties:
              Key:
                type: string
              Value:
                type: string
...
paths:
  /s3:
    get:
      summary: List S3 buckets
      operationId: listBuckets
      responses:
        '200':
          description: Successfully retrieved buckets
          content:
            application/json:
              schema:
                type: object
                properties:
                  buckets:
                    type: array
                    items:
                      $ref: '#/components/schemas/S3Bucket'
                  requestId:
                    type: string
                  timestamp:
                    type: string
                  region:
                    type: string
        '401':
          description: AWS Authentication Failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      summary: Create S3 bucket
      operationId: createBucket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/S3Bucket'
      responses:
        '201':
          description: Bucket created successfully
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: AWS Authentication Failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /lambda/functions:
    get:
      summary: List Lambda functions
      operationId: listFunctions
      responses:
        '200':
          description: Successfully retrieved Lambda functions
          content:
            application/json:
              schema:
                type: object
                properties:
                  functions:
                    type: array
                    items:
                      $ref: '#/components/schemas/LambdaFunction'
                  requestId:
                    type: string
                  timestamp:
                    type: string
        '401':
          description: AWS Authentication Failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      summary: Create Lambda function
      operationId: createFunction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LambdaFunction'
      responses:
        '201':
          description: Function created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LambdaFunction'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: AWS Authentication Failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /apigateway:
    get:
      summary: List API Gateway APIs
      operationId: listApis
      responses:
        '200':
          description: Successfully retrieved APIs
          content:
            application/json:
              schema:
                type: object
                properties:
                  apis:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApiGateway'
                  requestId:
                    type: string
                  timestamp:
                    type: string
        '401':
          description: AWS Authentication Failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      summary: Create API Gateway
      operationId: createApi
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiGateway'
      responses:
        '201':
          description: API Gateway created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiGateway'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: AWS Authentication Failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /iam/roles:
    get:
      summary: List IAM roles
      operationId: listRoles
      responses:
        '200':
          description: Successfully retrieved roles
          content:
            application/json:
              schema:
                type: object
                properties:
                  roles:
                    type: array
                    items:
                      $ref: '#/components/schemas/IAMRole'
                  requestId:
                    type: string
                  timestamp:
                    type: string
        '401':
          description: AWS Authentication Failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      summary: Create IAM role
      operationId: createRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IAMRole'
      responses:
        '201':
          description: Role created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IAMRole'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: AWS Authentication Failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /dynamodb/tables:
    get:
      summary: List DynamoDB tables
      operationId: listTables
      responses:
        '200':
          description: Successfully retrieved tables
          content:
            application/json:
              schema:
                type: object
                properties:
                  tables:
                    type: array
                    items:
                      type: string
                  requestId:
                    type: string
                  timestamp:
                    type: string
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalError'
    
    post:
      summary: Create DynamoDB table
      operationId: createTable
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DynamoDBTable'
      responses:
        '201':
          description: Table created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  tableName:
                    type: string
                  requestId:
                    type: string
                  timestamp:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalError'
...